remove the unused files and clean the work add userID to the pinecone using namespace
